fix: routing and user access
added UserDataController. Changed the login logic in the condition manager to redirect to admin.home. Changed userAccess and web.php middleware code to improve user access routing. Changed the link in the User Manager in the Sidebar. Fixed the location of the @endsection code in dashboard/books/index.blade.php. Added page for user-management.